db Dataset

This dataset contains the following variables:

operator

table for operators

project

table for projects

provider

table for (data) providers

action

table for action (e.g data download, data privider)

p

table for geographic/geospatial locations (e.g. stations/gauges locations), a sf object

variable

table for variaples (e.g. weather variales: precipitation/rainfall,temparature,...)

ts

table with time serirs (long table format)

attribute

table for attributes

data(db)

Format

Data frames and vectors

Source

GSOD (through GSODR package)

This dataset is intended for research purposes only, being distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY.

Details

This dataset stores all information about meteorological stations and instrumental timeseries. The user can easily use the package with their own data after replacing the values of such variables.

Examples

## see below

   library(nomnoml)
   data(db)
   str(db)
#> List of 8
#>  $ operator :Classes ‘data.table’ and 'data.frame':	2 obs. of  5 variables:
#>   ..$ ID       : int [1:2] 1 2
#>   ..$ name     : chr [1:2] "ecor" "jellow"
#>   ..$ firstname: chr [1:2] "Emanuele" "John"
#>   ..$ lastname : chr [1:2] "Cordano" "Field"
#>   ..$ email    : chr [1:2] "emanuele.cordano@gmail,com" NA
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ project  :Classes ‘data.table’ and 'data.frame':	3 obs. of  2 variables:
#>   ..$ ID  : int [1:3] 1 2 3
#>   ..$ name: chr [1:3] "rendena101" "rendena102" "rendena103"
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ provider :Classes ‘data.table’ and 'data.frame':	2 obs. of  2 variables:
#>   ..$ ID  : int [1:2] 1 2
#>   ..$ name: chr [1:2] "dwd" "gsod"
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ action   :Classes ‘data.table’ and 'data.frame':	3 obs. of  5 variables:
#>   ..$ ID         : int [1:3] 1 2 3
#>   ..$ description: chr [1:3] "download" "delivery" "download"
#>   ..$ timestamptz: POSIXct[1:3], format: "2025-03-17 11:41:27" "2025-03-17 11:41:27" ...
#>   ..$ operator_id: int [1:3] 1 1 1
#>   ..$ project_id : num [1:3] 1 1 1
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ p        :Classes ‘sf’ and 'data.frame':	122 obs. of  5 variables:
#>   ..$ ID         : int [1:122] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ name2      : chr [1:122] "GSOD_637560-99999 (MWANZA - TZ)" "GSOD_637900-99999 (MOSHI - TZ)" "GSOD_638010-99999 (KIGOMA - TZ)" "GSOD_638160-99999 (SAME - TZ)" ...
#>   ..$ description: chr [1:122] "GSOD station: data through GSODR R package , STNID : 637560-99999  , NAME : MWANZA  , CTRY : TZ  , COUNTRY : TA"| __truncated__ "GSOD station: data through GSODR R package , STNID : 637900-99999  , NAME : MOSHI  , CTRY : TZ  , COUNTRY : TAN"| __truncated__ "GSOD station: data through GSODR R package , STNID : 638010-99999  , NAME : KIGOMA  , CTRY : TZ  , COUNTRY : TA"| __truncated__ "GSOD station: data through GSODR R package , STNID : 638160-99999  , NAME : SAME  , CTRY : TZ  , COUNTRY : TANZ"| __truncated__ ...
#>   ..$ provider_id: num [1:122] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ geometry   :sfc_POINT of length 122; first list element:  'XY' num [1:2] 32.93 -2.44
#>   ..- attr(*, "sf_column")= chr "geometry"
#>   ..- attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA NA
#>   .. ..- attr(*, "names")= chr [1:4] "ID" "name2" "description" "provider_id"
#>  $ variable :Classes ‘data.table’ and 'data.frame':	21 obs. of  4 variables:
#>   ..$ ID         : int [1:21] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ name       : chr [1:21] "TEMP" "DEWP" "SLP" "STP" ...
#>   ..$ description: chr [1:21] "Mean daily temperature converted to degrees C to tenths. Missing = NA" "Mean daily dew point converted to degrees C to tenths. Missing = NA" "Mean sea level pressure in millibars to tenths. Missing = NA" "Mean station pressure for the day in millibars to tenths. Missing = NA" ...
#>   ..$ provider_id: num [1:21] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ ts       :Classes ‘data.table’ and 'data.frame':	7579551 obs. of  8 variables:
#>   ..$ ID          : int [1:7579551] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ attribute_id: int [1:7579551] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ variable_id : int [1:7579551] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ timestamptz : POSIXct[1:7579551], format: "1960-03-01" "1960-03-02" ...
#>   ..$ p_id        : int [1:7579551] 24 24 24 24 24 24 24 24 24 24 ...
#>   ..$ value       : num [1:7579551] 22.5 20.6 18.3 22.4 23.1 22.6 20.8 21.7 20.4 22.1 ...
#>   ..$ attrvalue   : chr [1:7579551] "4" "4" "4" "4" ...
#>   ..$ action_id   : num [1:7579551] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
#>  $ attribute:Classes ‘data.table’ and 'data.frame':	9 obs. of  4 variables:
#>   ..$ ID           : int [1:9] 1 2 3 4 5 6 7 8 9
#>   ..$ name         : chr [1:9] "TEMP_ATTRIBUTES" "DEWP_ATTRIBUTES" "SLP_ATTRIBUTES" "STP_ATTRIBUTES" ...
#>   ..$ description  : chr [1:9] "Number of observations used in calculating mean daily temperature" "Number of observations used in calculating mean daily dew point" "Number of observations used in calculating mean sea level pressure" "Number of observations used in calculating mean station pressure" ...
#>   ..$ variable_name: chr [1:9] "TEMP" "DEWP" "SLP" "STP" ...
#>   ..- attr(*, ".internal.selfref")=<externalptr> 
   schema <- lapply(db, FUN=function(t, nfks) {
    print(nfks)
    o <- sapply(as.list(t), FUN=function(x) {class(x)})
    o2 <- data.frame(column=names(o), type=as.character(o))
      o2$reference <- as.character(NA)
      iin <- which(o2$column %in% nfks)
      o2$reference[iin] <- o2$column[iin]
      o2$reference[iin] <- stringr::str_sub(o2$reference[iin], end=-4)
      return(o2)
    }, nfks=sprintf("%s_id", names(db)))
#> [1] "operator_id"  "project_id"   "provider_id"  "action_id"    "p_id"        
#> [6] "variable_id"  "ts_id"        "attribute_id"
#> [1] "operator_id"  "project_id"   "provider_id"  "action_id"    "p_id"        
#> [6] "variable_id"  "ts_id"        "attribute_id"
#> [1] "operator_id"  "project_id"   "provider_id"  "action_id"    "p_id"        
#> [6] "variable_id"  "ts_id"        "attribute_id"
#> [1] "operator_id"  "project_id"   "provider_id"  "action_id"    "p_id"        
#> [6] "variable_id"  "ts_id"        "attribute_id"
#> [1] "operator_id"  "project_id"   "provider_id"  "action_id"    "p_id"        
#> [6] "variable_id"  "ts_id"        "attribute_id"
#> [1] "operator_id"  "project_id"   "provider_id"  "action_id"    "p_id"        
#> [6] "variable_id"  "ts_id"        "attribute_id"
#> [1] "operator_id"  "project_id"   "provider_id"  "action_id"    "p_id"        
#> [6] "variable_id"  "ts_id"        "attribute_id"
#> [1] "operator_id"  "project_id"   "provider_id"  "action_id"    "p_id"        
#> [6] "variable_id"  "ts_id"        "attribute_id"

    out2 <- list()
    for (it in names(schema)) {
      print(it)
      content <- paste(schema[[it]]$column, schema[[it]]$type, sep = ": ", collapse = "||")
      out2[[it]] <- sprintf("[<table>%s| %s]", it, content)
      inn <- which(!is.na(schema[[it]]$reference))
      if (length(inn) > 0) { 
      refs <- sprintf("[%s] <- 1..n[%s]", schema[[it]]$reference[inn], it)
      out2[[it]] <- c(out2[[it]], refs)
    }
 }
#> [1] "operator"
#> [1] "project"
#> [1] "provider"
#> [1] "action"
#> [1] "p"
#> [1] "variable"
#> [1] "ts"
#> [1] "attribute"

 nomnoml(paste(unlist(out2), collapse="\n"), svg=TRUE)