Get station (or measurement points)

Get station (or measurement points)

get_p(x, p = NULL, p_table_name = "p", data.table = FALSE, sf = TRUE, ...)

Arguments

x

database or dateset, e.g like db

p

spatial (e.g. sf) object in which extract the observation station.

p_table_name

p table name

data.table, sf, ...

further arguments get_table

Examples

data(db)
p <- get_p(x=db)

library(terra)
#> terra 1.9.41
#> 
#> Attaching package: ‘terra’
#> The following object is masked from ‘package:data.table’:
#> 
#>     shift
library(leaflet)

plet(vect(p),cex=10)

### IN an area 

px <- p[49,] |> st_buffer(dist=250*1000) 
p1 <- get_p(x=db,p=px)
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries

plet(vect(p),cex=10)

px <- p[49,]
p1 <- get_p(x=db,p=px)
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries

plet(vect(p),cex=10)

north <- -2.25
south <- -8.85
west <- 29.45
east <- 32.15
pee <- rast(xmin=west,xmax=east,ymin=south,ymax=north,crs="EPSG:4326")
p2 <- get_p(x=db,p=pee)
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries