Get variables (or measurement types)

Get variables (or measurement types)

get_variable(x, variable_table_name = "variable", ...)

Arguments

x

database or dateset, e.g like db

variable_table_name

variable table name

...

further arguments

Examples

data(db)
variables <- get_variable(x=db)
#>        ID             name
#>     <int>           <char>
#>  1:     1             TEMP
#>  2:     2             DEWP
#>  3:     3              SLP
#>  4:     4              STP
#>  5:     5            VISIB
#>  6:     6             WDSP
#>  7:     7            MXSPD
#>  8:     8             GUST
#>  9:     9              MAX
#> 10:    10              MIN
#> 11:    11             PRCP
#> 12:    12             SNDP
#> 13:    13            I_FOG
#> 14:    14   I_RAIN_DRIZZLE
#> 15:    15       I_SNOW_ICE
#> 16:    16           I_HAIL
#> 17:    17        I_THUNDER
#> 18:    18 I_TORNADO_FUNNEL
#> 19:    19               EA
#> 20:    20               ES
#> 21:    21               RH
#>        ID             name
#>     <int>           <char>
#>                                                                                                                                                                                                                                                                                                  description
#>                                                                                                                                                                                                                                                                                                       <char>
#>  1:                                                                                                                                                                                                                                    Mean daily temperature converted to degrees C to tenths. Missing = NA
#>  2:                                                                                                                                                                                                                                      Mean daily dew point converted to degrees C to tenths. Missing = NA
#>  3:                                                                                                                                                                                                                                             Mean sea level pressure in millibars to tenths. Missing = NA
#>  4:                                                                                                                                                                                                                                   Mean station pressure for the day in millibars to tenths. Missing = NA
#>  5:                                                                                                                                                                                                                              Mean visibility for the day converted to kilometres to tenths. Missing = NA
#>  6:                                                                                                                                                                                                                           Mean daily wind speed value converted to metres/second to tenths. Missing = NA
#>  7:                                                                                                                                                                                                     Maximum sustained wind speed reported for the day converted to metres/second to tenths. Missing = NA
#>  8:                                                                                                                                                                                                                Maximum wind gust reported for the day converted to metres/second to tenths. Missing = NA
#>  9:                                                                                                Maximum temperature reported during the day converted to Celsius to tenths–time of max temp report varies by country and region, so this will sometimes not be the max for the calendar day. Missing = NA
#> 10:                                                                                                Minimum temperature reported during the day converted to Celsius to tenths–time of min temp report varies by country and region, so this will sometimes not be the max for the calendar day. Missing = NA
#> 11: Total precipitation (rain and/or melted snow) reported during the day converted to millimetres to hundredths; will usually not end with the midnight observation, i.e., may include latter part of previous day. A value of “.00” indicates no measurable precipitation (includes a trace). Missing = NA
#> 12:                                                                                                                                                                                                                                                        Snow depth in millimetres to tenths. Missing = NA
#> 13:                                                                                                                                                                                                                      Indicator for fog, (1 = yes, 0 = no/not reported) for the occurrence during the day
#> 14:                                                                                                                                                                                                          Indicator for rain or drizzle, (1 = yes, 0 = no/not reported) for the occurrence during the day
#> 15:                                                                                                                                                                                                      Indicator for snow or ice pellets, (1 = yes, 0 = no/not reported) for the occurrence during the day
#> 16:                                                                                                                                                                                                                     Indicator for hail, (1 = yes, 0 = no/not reported) for the occurrence during the day
#> 17:                                                                                                                                                                                                                  Indicator for thunder, (1 = yes, 0 = no/not reported) for the occurrence during the day
#> 18:                                                                                                                                                                                                  Indicator for tornado or funnel cloud, (1 = yes, 0 = no/not reported) for the occurrence during the day
#> 19:                                                                                                                                                              Mean daily actual vapour pressure as calculated using improved August-Roche-Magnus approximation (Alduchov and Eskridge 1996). Missing = NA
#> 20:                                                                                                                                                          Mean daily saturation vapour pressure as calculated using improved August-Roche-Magnus approximation (Alduchov and Eskridge 1996). Missing = NA
#> 21:                                                                                                                                                                   Mean daily relative humidity as calculated using improved August-Roche-Magnus approximation (Alduchov and Eskridge 1996). Missing = NA
#>                                                                                                                                                                                                                                                                                                  description
#>                                                                                                                                                                                                                                                                                                       <char>
#>     provider_id
#>           <num>
#>  1:           2
#>  2:           2
#>  3:           2
#>  4:           2
#>  5:           2
#>  6:           2
#>  7:           2
#>  8:           2
#>  9:           2
#> 10:           2
#> 11:           2
#> 12:           2
#> 13:           2
#> 14:           2
#> 15:           2
#> 16:           2
#> 17:           2
#> 18:           2
#> 19:           2
#> 20:           2
#> 21:           2
#>     provider_id
#>           <num>
variables <- get_variable(x=db,name="PRCP")
#>        ID             name
#>     <int>           <char>
#>  1:     1             TEMP
#>  2:     2             DEWP
#>  3:     3              SLP
#>  4:     4              STP
#>  5:     5            VISIB
#>  6:     6             WDSP
#>  7:     7            MXSPD
#>  8:     8             GUST
#>  9:     9              MAX
#> 10:    10              MIN
#> 11:    11             PRCP
#> 12:    12             SNDP
#> 13:    13            I_FOG
#> 14:    14   I_RAIN_DRIZZLE
#> 15:    15       I_SNOW_ICE
#> 16:    16           I_HAIL
#> 17:    17        I_THUNDER
#> 18:    18 I_TORNADO_FUNNEL
#> 19:    19               EA
#> 20:    20               ES
#> 21:    21               RH
#>        ID             name
#>     <int>           <char>
#>                                                                                                                                                                                                                                                                                                  description
#>                                                                                                                                                                                                                                                                                                       <char>
#>  1:                                                                                                                                                                                                                                    Mean daily temperature converted to degrees C to tenths. Missing = NA
#>  2:                                                                                                                                                                                                                                      Mean daily dew point converted to degrees C to tenths. Missing = NA
#>  3:                                                                                                                                                                                                                                             Mean sea level pressure in millibars to tenths. Missing = NA
#>  4:                                                                                                                                                                                                                                   Mean station pressure for the day in millibars to tenths. Missing = NA
#>  5:                                                                                                                                                                                                                              Mean visibility for the day converted to kilometres to tenths. Missing = NA
#>  6:                                                                                                                                                                                                                           Mean daily wind speed value converted to metres/second to tenths. Missing = NA
#>  7:                                                                                                                                                                                                     Maximum sustained wind speed reported for the day converted to metres/second to tenths. Missing = NA
#>  8:                                                                                                                                                                                                                Maximum wind gust reported for the day converted to metres/second to tenths. Missing = NA
#>  9:                                                                                                Maximum temperature reported during the day converted to Celsius to tenths–time of max temp report varies by country and region, so this will sometimes not be the max for the calendar day. Missing = NA
#> 10:                                                                                                Minimum temperature reported during the day converted to Celsius to tenths–time of min temp report varies by country and region, so this will sometimes not be the max for the calendar day. Missing = NA
#> 11: Total precipitation (rain and/or melted snow) reported during the day converted to millimetres to hundredths; will usually not end with the midnight observation, i.e., may include latter part of previous day. A value of “.00” indicates no measurable precipitation (includes a trace). Missing = NA
#> 12:                                                                                                                                                                                                                                                        Snow depth in millimetres to tenths. Missing = NA
#> 13:                                                                                                                                                                                                                      Indicator for fog, (1 = yes, 0 = no/not reported) for the occurrence during the day
#> 14:                                                                                                                                                                                                          Indicator for rain or drizzle, (1 = yes, 0 = no/not reported) for the occurrence during the day
#> 15:                                                                                                                                                                                                      Indicator for snow or ice pellets, (1 = yes, 0 = no/not reported) for the occurrence during the day
#> 16:                                                                                                                                                                                                                     Indicator for hail, (1 = yes, 0 = no/not reported) for the occurrence during the day
#> 17:                                                                                                                                                                                                                  Indicator for thunder, (1 = yes, 0 = no/not reported) for the occurrence during the day
#> 18:                                                                                                                                                                                                  Indicator for tornado or funnel cloud, (1 = yes, 0 = no/not reported) for the occurrence during the day
#> 19:                                                                                                                                                              Mean daily actual vapour pressure as calculated using improved August-Roche-Magnus approximation (Alduchov and Eskridge 1996). Missing = NA
#> 20:                                                                                                                                                          Mean daily saturation vapour pressure as calculated using improved August-Roche-Magnus approximation (Alduchov and Eskridge 1996). Missing = NA
#> 21:                                                                                                                                                                   Mean daily relative humidity as calculated using improved August-Roche-Magnus approximation (Alduchov and Eskridge 1996). Missing = NA
#>                                                                                                                                                                                                                                                                                                  description
#>                                                                                                                                                                                                                                                                                                       <char>
#>     provider_id
#>           <num>
#>  1:           2
#>  2:           2
#>  3:           2
#>  4:           2
#>  5:           2
#>  6:           2
#>  7:           2
#>  8:           2
#>  9:           2
#> 10:           2
#> 11:           2
#> 12:           2
#> 13:           2
#> 14:           2
#> 15:           2
#> 16:           2
#> 17:           2
#> 18:           2
#> 19:           2
#> 20:           2
#> 21:           2
#>     provider_id
#>           <num>
#> Warning: NAs introduced by coercion
#> [1] "name %in% c('PRCP')"